Fix custom JSON converter on enum (or other type)
System.Text.Json prefers converters defined in the Converters list,
and only then uses the JsonConverter attribute.
This is different from Newtonsoft, and does not work well with our
"catch all" converters which should be treated more like builtins
(with lowest priority).

The patchs add a converter which swaps the priority